Poorvika_PR1_Backend/requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
dj-rest-auth==2.2.4
mysqlclient==2.1.1
Django==4.0.8
django-appconf==1.0.5
django-cors-headers==3.12.0
django-filter==21.1
django-heroku==0.3.1
django-import-export==2.8.0
django-phonenumber-field==6.1.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-staticfiles==1.2.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.20.0
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.3
import-export==0.2.67.dev6
importlib-metadata==4.12.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.3.7
MarkupPy==1.14
MarkupSafe==2.1.1
modeltranslation==0.25
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.10
packaging==21.3
phonenumbers==8.12.46
psycopg2==2.9.3
PyJWT==2.4.0
pyparsing==3.0.9
pytz==2022.1
PyYAML==6.0
requests==2.28.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
whitenoise==6.2.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.8.0
